siakad/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
python-dotenv==1.0.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
marshmallow==3.21.2
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.3
